LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/fieldset-source-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_approx_equals: Unexpected start offset for input1 expected -74 +/- 0.1 but got -56.40625
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_approx_equals: Unexpected start offset for input2 expected -48 +/- 0.1 but got -30.40625
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_approx_equals: Unexpected start offset for input3 expected -22 +/- 0.1 but got -4.40625
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_approx_equals: Unexpected start offset for input4 expected 4 +/- 0.1 but got 21.59375
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_approx_equals: Unexpected start offset for input5 expected 30 +/- 0.1 but got 47.59375
Reservation Details
Name:

Date:

Time:

Number of guests:

Contact info:


PASS Fieldset is a valid source for a view timeline
